requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_version >= '3.7'
attrs==23.1.0 ; python_version >= '3.7'
charset-normalizer==3.3.0 ; python_full_version >= '3.7.0'
disnake==2.9.0
frozenlist==1.4.0 ; python_version >= '3.8'
idna==3.4 ; python_version >= '3.5'
multidict==6.0.4 ; python_version >= '3.7'
paho-mqtt==1.6.1
pure-cdb==4.0.0
pyyaml==6.0.1
yarl==1.9.2 ; python_version >= '3.7'
